fixing _Reader interface implementations to what's expected
in all #read family implementations, length is anything coerced to
integer, and the same applies to buffer (anything coercing to string).

One particular thing which I couldn't narrow down in the interface is
the following (quoting the docs):

> When this method is called at end of file, it returns nil or "",
> depending on length: read, read(nil), and read(0) return "",
> read(positive_integer) returns nil.

I suspect that this should also be defined in the API, as length will
rarely be 0 or nil in most practical cases.