(tentative) added support for inheriting from delegator classes
By defining such RBS classes such as:

```ruby
class Tempfile < Delegator[File]
...
```

with this patch, while building the instance definition, the class
assigned to the Delegator definition is then merged with the main
class definitions, which achieves the purpose of inheriting method
definitions from the delegated class.

This is a solution based on a stdlib definition however, so it might not
be the best way to move forward.