Add support for RSpec's verifying doubles in test setup helper
Besides regular doubles, RSpec also provides instance verifying doubles,
which are stricter and check that the stubbed methods actually do exist
in the original class or object.

In my experience, they tend to be preferred over regular doubles and
thus adding support for them is essential for RBS test hooks to be
useful out of the box, without specifying both the test suite and
additional classes to be ignored.

All classes listed in
https://github.com/rspec/rspec-mocks/blob/v3.11.1/lib/rspec/mocks/verifying_double.rb
have been added in this commit.

References:
- https://relishapp.com/rspec/rspec-mocks/docs/verifying-doubles