Improve String#scan type when the pattern is a String
String#scan's returned value is complex. It returns not only a `Array[String]` but also `Array[Array[String]]`. The returned value type depends on existence of Regexp's grouping. So the RBS user needs to assert the returned type manually.

This patch simplifies the returned type when the pattern argument is a
string. When the pattern is a string, it never contains Regexp's
grouping. It means it always returns a `Array[String]` when the argument
is a string.

By this change, the user does not need to care about the `Array[Array[String]]` case if the pattern is a string.