Implement multiple validate
**no option**

It will report as many validation errors as possible.
Warnings and errors are separated by log levels.
If there is even one error, the process will end with `exit 1`.

**--fail-fast**

If validation fails, it will be reported immediately and the process will end.
This is similar to the traditional individual validation.

**--exit-error-on-syntax-error**

It will report all errors, including syntax errors,
but if there is even one syntax error, the process will terminate with `exit 1`.

**--fail-fast --exit-error-on-syntax-error**

Even syntax errors, which would not normally be considered failures,
will be treated as errors and the process will immediately terminate with `exit 1`.