Correctly convert camel case to snake case when generating names in C files (#32)
Currently, camel case names in `config.yml`, like `MethodDefinition`,
will be generated as `methoddefinition`, which is not ideal. So this PR
updates `template.rb` to make sure they're generated as
`method_definition` instead.

I also changed `rbs_typename` to `rbs_type_name` as I think it's also
not ideal.

NO Ruby files are touched so it shouldn't be breaking for end users.