Allows the use of a path list as RBS_SKIP_TESTS
In ruby/ruby test-bundled-gems, rbs has many failures on Windows.
Currently we ignore all failures in this test totally, but that is the
test-run makes no sense but just wastes times.  So I want to also
specify another file for the tests to be skipped only on Windows, in
addition to the platform independent skipping list file, that we are
using already.